// 04-noticias/src/app/interfaces/interfaces.ts
export interface Source {
    id: string;
    name: string;
}

export interface Article {
    source: Source;
    author: string;
    title: string;
    description: string;
    url: string;
    urlToImage: string;
    publishedAt: Date;
    content: string;
}

export interface RespuestaTopHeadlines {
    status: string;
    totalResults: number;
    articles: Article[];
}